Moved more resin code to shared
